fix:#263 NotificationControllerのmarkAsReadメソッドを分離し、Apiフォルダに作成したNotificationController内に配置しました
